moved dateSortopions to its own component file
